requirements.txt
requests
websockets
protobuf
python-dotenv
